removed categories table, edited controllers, models, serializers

+++ app/controllers/projects_controller.rb
@@ -1,17 +1,18 @@
 class ProjectsController < ApplicationController
     rescue_from ActiveRecord::RecordNotFound, with: :render_not_found_response
 
-    before_action :confirm_admin, only: [:update, :destroy]
+    skip_before_action :authorize, only: [:index, :projects_by_category, :show]
    
-    #  'categories/:id/projects' OR '/projects' displays all projects or all projects of a certain category to user with pagination 
+    #  '/projects' displays all projects to visitors and users with pagination 
     def index
-        if params[:category_id]
-            category = Category.find(params[:category_id])
-            projects = category.projects
-        else
-            projects = Project.all
-        end
-        render json: projects.paginate(page: params[:page], per_page: 20), status: :ok
+        projects = Project.all
+        render json: projects.paginate(page: params[:page], per_page: 50), status: :ok
+    end
+
+    # '/:category/projects' displays all projects within the specified category to visitors and users with pagination 
+    def projects_by_category
+        projects = Project.where("category = ?", params[:category])
+        render json: projects.paginate(page: params[:page], per_page: 50), status: :ok
     end
 
     #  'projects/:id' displays individual project 
@@ -23,20 +24,24 @@ class ProjectsController < ApplicationController
     # '/projects' lets user create a project, automatically adds project to their user_projects
     def create
         project = Project.create!(project_params)
+        project.shared_by = @current_user.username
+        project.save
         UserProject.create!(user_id: @current_user.id, project_id: project.id, completed_status: user_project_params[:completed_status])
         render json: project, status: :created
     end
 
-    # '/projects/:id' ADMIN ONLY - updates a project
+    # '/projects/:id' lets users edit only the projects they have shared
     def update
         project = find_project
+        return render json: {error: "Projects can only be edited or deleted by the user who orginally shared them" }, status: :unauthorized unless project.shared_by == @current_user.username
         project.update!(project_params)
         render json: project, status: :ok
     end
 
-    # '/projects/:id' ADMIN ONLY - deletes a project and deletes associated user_projects
+    # '/projects/:id' lets users delete only the projects they have shared 
     def destroy
         project = find_project
+        return render json: {error: "Projects can only be edited or deleted by the user who orginally shared them" }, status: :unauthorized unless project.shared_by == @current_user.username
         project.destroy
         head :no_content
     end
@@ -48,7 +53,7 @@ class ProjectsController < ApplicationController
     end
 
     def project_params
-        params.permit(:title, :description, :url, :image, :category_id)
+        params.permit(:title, :description, :url, :image, :category)
     end
 
     def user_project_params
@@ -59,7 +64,8 @@ class ProjectsController < ApplicationController
         render json: {error: "Project not found" }, status: :not_found
     end
 
-    def confirm_admin
-        return render json: {error: "Projects can only be deleted and edited by admin" }, status: :unauthorized unless @current_user.admin
-    end
+    # def confirm_admin
+    #     return render json: {error: "Projects can only be deleted and edited by admin" }, status: :unauthorized unless @current_user.admin
+    # end
 end
+
